Use null rather than empty string for unspecified namespace
It's possible for (arguably misformed) package metadata to have the
empty string as the namespace. This can happen for example when
ingesting an OCI image that does not have the correct repository URL
specified.

Currently guac-visualiser does not let you pick beyond a namespace
that is the empty string: it would not make the "package name"
dropdown sensitive (although it would correctly populate the options
from the GraphQL response).

This commit updates the "empty" state for the package name option from
the empty string to null (like is used for version). Then update the
`disabled` param to be determined on whether the namespace is null
(rather than the empty string).

Signed-off-by: Ed Baunton <edbaunton@gmail.com>